前端使用了vue-element-admin模板，并进行了二次开发，目前完成了菜单管理和角色管理（其中，角色管理的权限分配尚未完成），后端继续完善了权限管理系统的功能，并重新提交了周报和监督周报

+++ backend/src/main/java/Exmpl/Entity/Role.java
@@ -9,7 +9,7 @@ import lombok.Data;
 @TableName("role")
 public class Role {
     @TableId
-    private int rno;
+    private Long rno;
     private String rolename;
     private boolean  isAdmin;
 //    private boolean isUpdate = false;
